requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
dnspython==2.4.2
jinja2==3.1.2
pydantic==2.4.2
python-whois==0.8.0
